+++ core/Lucy/Search/PhraseQuery.c
@@ -68,7 +68,7 @@ static PhraseQuery*
 S_do_init(PhraseQuery *self, String *field, Vector *terms, float boost) {
     Query_init((Query*)self, boost);
     PhraseQueryIVARS *const ivars = PhraseQuery_IVARS(self);
-    for (uint32_t i = 0, max = Vec_Get_Size(terms); i < max; i++) {
+    for (size_t i = 0, max = Vec_Get_Size(terms); i < max; i++) {
         CERTIFY(Vec_Fetch(terms, i), OBJ);
     }
     ivars->field = field;
@@ -138,7 +138,7 @@ PhraseQuery_Equals_IMP(PhraseQuery *self, Obj *other) {
 String*
 PhraseQuery_To_String_IMP(PhraseQuery *self) {
     PhraseQueryIVARS *const ivars = PhraseQuery_IVARS(self);
-    uint32_t  num_terms = Vec_Get_Size(ivars->terms);
+    uint32_t  num_terms = (uint32_t)Vec_Get_Size(ivars->terms);
     CharBuf  *buf       = CB_new(0);
     CB_Cat(buf, ivars->field);
     CB_Cat_Trusted_Utf8(buf, ":\"", 2);
@@ -218,7 +218,7 @@ PhraseCompiler_init(PhraseCompiler *self, PhraseQuery *parent,
 
     // Store IDF for the phrase.
     ivars->idf = 0;
-    for (uint32_t i = 0, max = Vec_Get_Size(terms); i < max; i++) {
+    for (size_t i = 0, max = Vec_Get_Size(terms); i < max; i++) {
         Obj     *term     = Vec_Fetch(terms, i);
         int32_t  doc_max  = Searcher_Doc_Max(searcher);
         int32_t  doc_freq = Searcher_Doc_Freq(searcher, parent_ivars->field, term);
@@ -300,7 +300,7 @@ PhraseCompiler_Make_Matcher_IMP(PhraseCompiler *self, SegReader *reader,
     PhraseQueryIVARS *const parent_ivars
         = PhraseQuery_IVARS((PhraseQuery*)ivars->parent);
     Vector *const      terms     = parent_ivars->terms;
-    uint32_t           num_terms = Vec_Get_Size(terms);
+    uint32_t           num_terms = (uint32_t)Vec_Get_Size(terms);
 
     // Bail if there are no terms.
     if (!num_terms) { return NULL; }
@@ -350,7 +350,7 @@ PhraseCompiler_Highlight_Spans_IMP(PhraseCompiler *self, Searcher *searcher,
         = PhraseQuery_IVARS((PhraseQuery*)ivars->parent);
     Vector *const      terms     = parent_ivars->terms;
     Vector *const      spans     = Vec_new(0);
-    const uint32_t     num_terms = Vec_Get_Size(terms);
+    const uint32_t     num_terms = (uint32_t)Vec_Get_Size(terms);
     UNUSED_VAR(searcher);
 
     // Bail if no terms or field doesn't match.
@@ -395,7 +395,7 @@ PhraseCompiler_Highlight_Spans_IMP(PhraseCompiler *self, Searcher *searcher,
     }
 
     // Proceed only if all terms are present.
-    uint32_t num_tvs = Vec_Get_Size(term_vectors);
+    uint32_t num_tvs = (uint32_t)Vec_Get_Size(term_vectors);
     if (num_tvs == num_terms) {
         TermVector *first_tv = (TermVector*)Vec_Fetch(term_vectors, 0);
         TermVector *last_tv
