+++ core/Lucy/Test/Analysis/TestStandardTokenizer.c
@@ -112,7 +112,7 @@ test_tokenizer(TestBatchRunner *runner) {
         Vector *tests = (Vector*)Json_slurp_json((Folder*)modules_folder, path);
         if (!tests) { RETHROW(Err_get_error()); }
 
-        for (uint32_t i = 0, max = Vec_Get_Size(tests); i < max; i++) {
+        for (size_t i = 0, max = Vec_Get_Size(tests); i < max; i++) {
             Hash *test = (Hash*)Vec_Fetch(tests, i);
             String *text = (String*)Hash_Fetch_Utf8(test, "text", 4);
             Vector *wanted = (Vector*)Hash_Fetch_Utf8(test, "words", 5);
